Use the textureMatrix supplied from SurfaceTexture
This fixes a bug where on some devices the camera preview wouldn’t appear correctly. In my case it had a green line a few pixels wide along the bottom edge of the preview.

Also, in my instance the texture matrix flipped the image, so I had to enable bFlipTexture. I don’t know if this is device dependent or not. I tested on two separate devices, a Saumsung Tab 2 and Nexus 5X.